#14: Move rtm listener classes to a subdirectory
