+++ src/bashi/conversation/instant-message-replier/instant-message-listener.ts
@@ -1,4 +1,4 @@
-import { IRtmListener, ISlackConnectionManager } from "../../interfaces";
+import { IRtmListener, ISlackConnectionManager } from "../../../interfaces";
 
 export class InstantMessageListener implements IRtmListener {
     private slackConnectionManager: ISlackConnectionManager;
