+++ src/interfaces/conversation/train-time-notifier/ITrainData.ts
