+++ src/interfaces/conversation/train-time-notifier/ITrainDataCrawler.ts
@@ -1,4 +1,4 @@
-import { ITrainData } from "../";
+import { ITrainData } from "../../";
 
 export interface ITrainDataCrawler {
     getTrainData(url): Promise<ITrainData[]>;
